+++ monkay/__init__.py
@@ -2,12 +2,20 @@
 #
 # SPDX-License-Identifier: BSD-3-Clauses
 
-from .base import DeprecatedImport, ExtensionProtocol, Monkay, load, load_any
+from .base import (
+    PRE_ADD_LAZY_IMPORT_HOOK,
+    DeprecatedImport,
+    ExtensionProtocol,
+    Monkay,
+    load,
+    load_any,
+)
 
 __all__ = [
     "Monkay",
     "SortedExportsEntry",
     "DeprecatedImport",
+    "PRE_ADD_LAZY_IMPORT_HOOK",
     "ExtensionProtocol",
     "load",
     "load_any",